refactor: Move toggle logic out of menu
