+++ apps/docs/components/side-menu/index.tsx
@@ -1,17 +1,9 @@
-import { useCallback, useEffect, useState } from 'react'
+import { useCallback, useState } from 'react'
 import { withRouter, NextRouter } from 'next/router'
-import { useTheme } from 'next-themes'
 
-import { log } from '@aviato/utils'
-import {
-  SideMenu,
-  Menu,
-  MenuItem,
-  Button,
-  styled,
-  getCurrentTheme,
-} from '@aviato/ui'
+import { SideMenu, Menu, MenuItem, styled } from '@aviato/ui'
 import { AviatoLogo } from '../logo'
+import { ToggleThemeButton } from '../toggle-theme'
 
 const mainMenu: MenuDataItems[] = [
   {
@@ -74,12 +66,6 @@ type MenuDataItems = {
 
 const MainSideMenu = withRouter(({ router }: MainSideMenuProps) => {
   const [activeRoute, setActiveRoute] = useState('/')
-  const [mounted, setMounted] = useState(false)
-  const { setTheme, theme } = useTheme()
-  const [currentTheme, setCurrentTheme] = useState(theme)
-
-  useEffect(() => setMounted(true), [])
-
   const { asPath } = router
 
   useCallback(() => {
@@ -95,20 +81,6 @@ const MainSideMenu = withRouter(({ router }: MainSideMenuProps) => {
     router.push({
       pathname: targetRoute,
     })
-  }
-
-  const toggleTheme = useCallback(() => {
-    log.global.debug('Toggling theme...')
-
-    const targetTheme = getCurrentTheme() === 'light' ? 'dark' : 'light'
-
-    setTheme(targetTheme)
-    setCurrentTheme(targetTheme)
-  }, [setTheme])
-
-  // Prevent SSR de-sync error by waiting for CSR
-  if (!mounted) {
-    return null
   }
 
   const mainMenuItems = mainMenu.map(({ title, route, subMenu }, menuIndex) => {
@@ -147,9 +119,7 @@ const MainSideMenu = withRouter(({ router }: MainSideMenuProps) => {
       <Menu>{mainMenuItems}</Menu>
 
       <Footer>
-        <Button onClick={() => toggleTheme()}>
-          {currentTheme === 'light' ? '🌙' : '☀️'}
-        </Button>
+        <ToggleThemeButton />
       </Footer>
     </SideMenu>
   )
